requirements.txt
aenum==3.1.0
aioredis==1.3.1
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dbf==0.99.1
dbfread==2.0.7
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-redis==5.1.0
django-timezone-field==4.2.1
djangorestframework==3.13.1
docxcompose==1.3.3
docxtpl==0.14.2
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Jinja2==3.0.3
kombu==5.1.0
lxml==4.7.1
MarkupSafe==2.0.1
msgpack==1.0.2
numpy==1.21.5
openpyxl==3.0.7
pandas==1.3.5
promise==2.3
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
redis==3.5.3
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
zope.interface==5.4.0
